changing back to direct loading of 311 and 911 data to map
