+++ experiment-PIT/public/data/process_311.ts
@@ -1,13 +1,10 @@
-import { getShotsData } from '../../src/api/api.ts';
-import { check } from "@placemarkio/check-geojson";
-
-//TO USE:
-//run this file on map load and download geojson file to give to mapbox to create vector tileset.
+import { get311Data } from '../../src/api/api.ts';
 
 interface GeoJSONFeature {
     type: string,
     properties: {
         id: number;
+        request_type: string,
         date: string;
     };
     geometry: {
@@ -16,60 +13,42 @@ interface GeoJSONFeature {
     }
 }
 
-//process shots data from api and turning into geojson
-export const processShotsData = async () => {
-    //loading 
-    const shots_data = await getShotsData();
-    const shots_geojson = { type: "FeatureCollection", features: [] as GeoJSONFeature[] }; //defining type of array
-
-    //converting to GeoJSON
-    for (const instance of shots_data){ //using for of instead of for in
-        const shot_id = instance.id;
-        const shot_latitude = instance.latitude;
-        const shot_longitude = instance.longitude;
-        const shot_date = instance.date;
-        //const shot_ballistics = instance.ballistics_evidence
-        //include ballistics evidence?
-
-        shots_geojson.features.push({
-            "type": "Feature",
-            "properties": {
-                id: shot_id,
-                date: shot_date,
-                //ballistics: shot_ballistics,
-            },
-            "geometry": {
-                "type": "Point",
-                "coordinates": [
-                    parseFloat(shot_longitude),
-                    parseFloat(shot_latitude)
-                ]
-            } 
-        })
-    }
-    
-    //
-    
-    // Create a downloadable file to export to give to mapbox
+export const process311Data = async () => {
     try {
-        const checkingObj = check(JSON.stringify(shots_geojson, null, 2));
-        console.log("validating geojson..");
-        console.log(checkingObj);
+         //loading 
+        const request_data = await get311Data();
+        const request_geojson = { type: "FeatureCollection", features: [] as GeoJSONFeature[] }; //defining type of array
 
-        const blob = new Blob([JSON.stringify(shots_geojson, null, 2)], { type: 'application/json' });
-        console.log("✅ blob created!")
-        const url = URL.createObjectURL(blob);
-    
-        const a = document.createElement('a');
-        a.href = url;
-        a.download = 'shots_data.geojson';
-        a.click();
-    
-        console.log('✅ GeoJSON file download triggered!');
+        //converting to geojson
+        for (const instance of request_data){
+            const request_id = instance.id;
+            const request_type = instance.type;
+            const request_latitude = instance.latitude;
+            const request_longitude = instance.longitude;
+            const request_date = instance.date;
+
+            request_geojson.features.push({
+                "type": "Feature",
+                "properties": {
+                    id: request_id,
+                    request_type: request_type,
+                    date: request_date,
+                },
+                "geometry": {
+                    "type": "Point",
+                    "coordinates": [
+                        parseFloat(request_latitude),
+                        parseFloat(request_longitude)
+                    ]
+                } 
+            })
+
+        }
+
+        return request_geojson;
+
     } catch (error) {
-        console.log('❌ Error validating, creating, or downloading the GeoJSON file:', error);
+        console.log("❌ Error loading 311 data from database or converting to geojson..", error);
     }
-   
+       
 }
-
-// or could turn into csv file to give to mapboxs
